Refactoring browser script to use unittest
Moved the code into a class in line with the unittest system.

The script now also loops through each video twice. We do this because
the video files are having playback problems in Chrome on the 2nd time
through the videos play fine.

I also simplified the Xpath to the various click and video buttons